requirements.txt
Flask>=3.0,<4
SQLAlchemy>=2.0
marshmallow>=3.20,<4
python-dotenv>=1.0,<2
gunicorn>=21,<23
requests>=2.31,<3
tqdm>=4.66,<5
psycopg2-binary>=2.9,<3
psycopg[binary]>=3.1,<4
